+++ lib/template_class.rb
@@ -1,9 +1,2 @@
-# frozen_string_literal: true
-
 require_relative 'template_class/version'
-
-module TemplateClass
-  class Error < StandardError; end
-end
-
 require_relative 'template_class/template'