+++ CHANGELOG.md
@@ -8,6 +8,17 @@
 ### Bug fixes
 )-->
 
+## 1.1.0 2023-08-31
+
+### New features
+
+- Added support for caching modules.
+- Added RBS type signatures.
+
+### Bug fixes
+
+- Removed the `Error` class.
+
 ## 1.0.0 2023-04-07
 
 First release. Refer to [README.md](README.md) for the full documentation.